.ide/Builders/CppBuilder.cpp: fixed my own (harmless) bug - BRC C is now generated using unsigned bytes
git-svn-id: svn://ultimatepp.org/upp/trunk@3938 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/ide/Builders/CppBuilder.cpp
@@ -372,7 +372,7 @@ static void WriteByteArray(StringBuffer& fo, const String& data)
 		}
 		if(pos == 0)
 			fo << '\t';
-		String part = FormatInt(data[p]);
+		String part = FormatInt((byte)data[p]);
 		fo << part << ", ";
 		pos += part.GetLength() + 2;
 	}
